requirements.txt
python-igraph==0.8.3
texttable==1.6.3
